+++ packages/service-core/src/storage/storage-index.ts
@@ -5,6 +5,7 @@ export * from './SourceEntity.js';
 export * from './SourceTable.js';
 export * from './StorageEngine.js';
 
+export * from './mongo/config.js';
 export * from './mongo/db.js';
 export * from './mongo/models.js';
 export * from './mongo/MongoBucketBatch.js';
@@ -17,5 +18,4 @@ export * from './mongo/MongoSyncRulesLock.js';
 export * from './mongo/OperationBatch.js';
 export * from './mongo/PersistedBatch.js';
 export * from './mongo/util.js';
-export * from './mongo/config.js';
-export * from './write-checkpoint.js';
+export * from './WriteCheckpointAPI.js';
